49: Fixing structure of the project
*Brief summary*
There was a problem how should we segrgate domains. It stood out that we
need to put checks, suites, aggregated checks and reports as a one huge
subdomain (for now).

*How I did it changes*
Moved repositories to infrastucture. All domain objects are placed
together.

*Additional changes*
Fix: Agregated Reports calls are not done through application layer